Make issue extraction concurrent
Since all issues are independent of one another,
we can use [`FuturesUnordered`](https://docs.rs/futures/latest/futures/stream/struct.FuturesUnordered.html)
to make issue search concurrent. This results in a nice speed bump.

We use a `Semaphore` to limit the concurrency to a maximum of
100 concurrent checks.
Note that we need to convert the `extract_link_and_title` and
`parse_page` methods to static members, as we'd otherwise run into
ownership issues with `self`. Fortunately, these methods were
independent of `self` anyway, so they should probably have been static
to begin with.

A similar optimization can be done for the lychee link check, but I'd
change that in a separate PR.